requirements.txt
Django==2.2.12
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
gunicorn==20.0.4